+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ## 1.17.8 - Unreleased
 
+* Fixed [#652](https://github.com/NeoVintageous/NeoVintageous/issues/652): Command-line output should all use the same output panel
+* Fixed [#651](https://github.com/NeoVintageous/NeoVintageous/issues/651): Command-line output should be read-only
 * Fixed [#650](https://github.com/NeoVintageous/NeoVintageous/issues/650): `C-d`, print the cursor position, is incorrect
 * Fixed [#649](https://github.com/NeoVintageous/NeoVintageous/issues/649): Enter visual block from visual selection is broken (ST4)
 * Fixed [#647](https://github.com/NeoVintageous/NeoVintageous/issues/647): Replace character bug (ST4)